Refactor to reuse GenericJUnitTestEventAdapter
