+++ subprojects/testing-jvm/src/main/java/org/gradle/api/internal/tasks/testing/junit/JUnitTestEventAdapter.java
@@ -17,90 +17,43 @@
 package org.gradle.api.internal.tasks.testing.junit;
 
 import org.gradle.api.internal.tasks.testing.DefaultTestDescriptor;
-import org.gradle.api.internal.tasks.testing.TestCompleteEvent;
 import org.gradle.api.internal.tasks.testing.TestDescriptorInternal;
 import org.gradle.api.internal.tasks.testing.TestResultProcessor;
-import org.gradle.api.internal.tasks.testing.TestStartEvent;
-import org.gradle.api.tasks.testing.TestResult;
 import org.gradle.internal.id.IdGenerator;
 import org.gradle.internal.time.Clock;
 import org.junit.runner.Description;
 import org.junit.runner.notification.Failure;
 import org.junit.runner.notification.RunListener;
 
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class JUnitTestEventAdapter extends RunListener {
     private static final Pattern DESCRIPTOR_PATTERN = Pattern.compile("(.*)\\((.*)\\)", Pattern.DOTALL);
-    protected final IdGenerator<?> idGenerator;
-    private final TestResultProcessor resultProcessor;
-    private final Clock clock;
-    private final Object lock = new Object();
-    private final Map<Object, TestDescriptorInternal> executing = new HashMap<Object, TestDescriptorInternal>();
-    private final Set<Object> assumptionFailed = new HashSet<Object>();
+    private final IdGenerator<?> idGenerator;
+    private final GenericJUnitTestEventAdapter<Description> adapter;
 
     public JUnitTestEventAdapter(TestResultProcessor resultProcessor, Clock clock,
                                  IdGenerator<?> idGenerator) {
-        assert resultProcessor instanceof org.gradle.internal.concurrent.ThreadSafe;
-        this.resultProcessor = resultProcessor;
-        this.clock = clock;
         this.idGenerator = idGenerator;
+        this.adapter = new GenericJUnitTestEventAdapter<Description>(resultProcessor, clock);
     }
 
     @Override
     public void testStarted(Description description) throws Exception {
-        testStarted(description, nullSafeDescriptor(idGenerator.generateId(), description));
-    }
-
-    protected void testStarted(Object identifier, TestDescriptorInternal descriptor) {
-        synchronized (lock) {
-            TestDescriptorInternal oldTest = executing.put(identifier, descriptor);
-            assert oldTest == null : String.format("Unexpected start event for %s", identifier);
-        }
-        resultProcessor.started(descriptor, startEvent());
+        adapter.testStarted(description, nullSafeDescriptor(idGenerator.generateId(), description));
     }
 
     @Override
     public void testFailure(Failure failure) throws Exception {
-        testFailure(failure.getDescription(), nullSafeDescriptor(idGenerator.generateId(), failure.getDescription()), failure.getException());
-    }
-
-    protected void testFailure(Object identifier, TestDescriptorInternal descriptor, Throwable exception) {
-        if (descriptor == null) {
-            return;
-        }
-        TestDescriptorInternal testInternal;
-        synchronized (lock) {
-            testInternal = executing.get(identifier);
-        }
-        boolean needEndEvent = false;
-        if (testInternal == null) {
-            // This can happen when, for example, a @BeforeClass or @AfterClass method fails
-            needEndEvent = true;
-            testInternal = descriptor;
-            resultProcessor.started(testInternal, startEvent());
-        }
-        resultProcessor.failure(testInternal.getId(), exception);
-        if (needEndEvent) {
-            resultProcessor.completed(testInternal.getId(), new TestCompleteEvent(clock.getCurrentTime()));
-        }
+        adapter.testFailure(failure.getDescription(), nullSafeDescriptor(idGenerator.generateId(), failure.getDescription()), failure.getException());
     }
 
     @Override
     public void testAssumptionFailure(Failure failure) {
-        testAssumptionFailure(failure.getDescription());
+        adapter.testAssumptionFailure(failure.getDescription());
     }
 
-    protected void testAssumptionFailure(Object identifier){
-        synchronized (lock) {
-            assumptionFailed.add(identifier);
-        }
-    }
 
     @Override
     public void testIgnored(Description description) throws Exception {
@@ -109,13 +62,8 @@ public class JUnitTestEventAdapter extends RunListener {
             // generate them on our own
             processIgnoredClass(description);
         } else {
-            testIgnored(descriptor(idGenerator.generateId(), description));
+            adapter.testIgnored(descriptor(idGenerator.generateId(), description));
         }
-    }
-
-    protected void testIgnored(TestDescriptorInternal descriptor) {
-        resultProcessor.started(descriptor, startEvent());
-        resultProcessor.completed(descriptor.getId(), new TestCompleteEvent(clock.getCurrentTime(), TestResult.ResultType.SKIPPED));
     }
 
     private void processIgnoredClass(Description description) throws Exception {
@@ -128,28 +76,7 @@ public class JUnitTestEventAdapter extends RunListener {
 
     @Override
     public void testFinished(Description description) throws Exception {
-        testFinished((Object) description);
-    }
-
-    protected void testFinished(Object identifier) {
-        long endTime = clock.getCurrentTime();
-        TestDescriptorInternal testInternal;
-        TestResult.ResultType resultType;
-        synchronized (lock) {
-            testInternal = executing.remove(identifier);
-            if (testInternal == null && executing.size() == 1) {
-                // Assume that test has renamed itself (this can actually happen)
-                testInternal = executing.values().iterator().next();
-                executing.clear();
-            }
-            assert testInternal != null : String.format("Unexpected end event for %s", identifier);
-            resultType = assumptionFailed.remove(identifier) ? TestResult.ResultType.SKIPPED : null;
-        }
-        resultProcessor.completed(testInternal.getId(), new TestCompleteEvent(endTime, resultType));
-    }
-
-    private TestStartEvent startEvent() {
-        return new TestStartEvent(clock.getCurrentTime());
+        adapter.testFinished(description);
     }
 
     private TestDescriptorInternal descriptor(Object id, Description description) {
